Grabbed Doug Lea's malloc.  It's *tons* better (in memory usage) than MSVC's malloc -- to the tune of 3x better when running chunkdemo against the puget data.  Keeps memory profile down to around 100MB, as opposed to the 300+ MB I was seeing before!
Usage of dl's malloc is controlled by a compile-time switch; at the moment it's not turned on by default except when using MSVC; hopefully most Linuxes (and Cygwin for that matter) will have taken advantage of dl's malloc and so there's no point in trying to replace it.  Investigate later.

Fixed a mip-mapping bug -- image::make_next_mip_level() needed to keep the pitch as a multiple of 4, even with 2x2 and 3x3 bitmaps.

Stopped using SDL_Surface* in chunklod; better to just use image::rgb and thus have true control over alloc/free of the image buffer (and be able to replace with dlmalloc/dlfree).