big refactor -- separate movies into instance and definition.
We want to be able to play several independent copies of one
movie.

Unresolved: sprite instances are yet another aspect of this
(a single movie instance can have multiple sprite instances), but
the instancing of movie characters is not completely consistent --
we would also like things like mutable text characters to be
instancable on a per-movie basis rather than per-display-object,
differently from how sprites
are handled.